backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
dj-database-url>=2.1,<3.0
psycopg[binary]>=3.1,<4.0
